% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSampleFeatures.R
\name{addsampleFeatures}
\alias{addsampleFeatures}
\title{addsampleFeatures}
\usage{
addsampleFeatures(object, sample.data = NULL, id.col = "sample")
}
\arguments{
\item{object}{cignatures class object}

\item{sample.data}{data.frame containing sample-level variables}

\item{id.col}{column containing sample identifiers}
}
\value{
cignaturesCN or cignaturesSIG object with updated samplefeatData
}
\description{
Adds custom sample-level data to the samplefeatData field of a cignatures object.
This can be additional sample information (purity, tumour type, etc.) that can
be used in downstream analysis.
}
