% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calculateFeatures.R
\docType{methods}
\name{calculateFeatures}
\alias{calculateFeatures}
\alias{calculateFeatures,cignaturesCN-method}
\title{calculateFeatures}
\usage{
calculateFeatures(object, method = "drews", smooth.diploid = TRUE)

\S4method{calculateFeatures}{cignaturesCN}(object, method = NULL, smooth.diploid = TRUE)
}
\arguments{
\item{object}{cignaturesCN object}

\item{method}{Method to extract copy number features. Default is "drews".}

\item{smooth.diploid}{Binary variable indicating whether segments close to 2 should be collapsed to 2 and merged together. Default is TRUE.}
}
\value{
A cignaturesCN class object with extracted features stored in the "featData" slot
}
\description{
Extracts and returns copy number features from copy number profiles in a cignaturesCN object.
}
