% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cignatures.R
\docType{package}
\name{cignatures}
\alias{cignatures}
\title{cignatures: A package for computating the notorious bar statistic}
\description{
The cignatures package provides a suite of tools for generating, analysing
and visualising copy number signatures
}
\section{Foo functions}{

The foo functions ...
}

