% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{cignaturesCN-class}
\alias{cignaturesCN-class}
\alias{cignaturesCN}
\title{cignaturesCN object}
\description{
cignaturesCN object
}
\section{Slots}{

\describe{
\item{\code{segments}}{list}

\item{\code{featData}}{list}

\item{\code{featFitting}}{list}

\item{\code{samplefeatData}}{data.frame}

\item{\code{ExpData}}{cignaturesExpData}
}}

