% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{cignaturesExpData-class}
\alias{cignaturesExpData-class}
\alias{cignaturesExpData}
\title{cignaturesExpData object}
\description{
cignaturesExpData object
}
\section{Slots}{

\describe{
\item{\code{experimentName}}{character}

\item{\code{init.date}}{character}

\item{\code{last.modified}}{character}

\item{\code{samples.full}}{numeric}

\item{\code{samples.current}}{numeric}

\item{\code{build}}{character}

\item{\code{feature.method}}{character}
}}

