% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/clinPredictionPlatinum.R
\docType{methods}
\name{clinPredictionPlatinum}
\alias{clinPredictionPlatinum}
\alias{clinPredictionPlatinum,cignaturesSIG-method}
\title{clinPredictionPlatinum}
\usage{
clinPredictionPlatinum(object)

\S4method{clinPredictionPlatinum}{cignaturesSIG}(object)
}
\arguments{
\item{object}{cignaturesSIG object}
}
\value{
A vector with "Predicted sensitive" or "Predicted resistant" for all samples in the input object.
}
\description{
The function takes signature activities based on Drews et al. methodology and predicts patient's response to platinum-based chemotherapies.
}
