% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getExperiment.R
\docType{methods}
\name{getExperiment}
\alias{getExperiment}
\alias{getExperiment,cignaturesCN-method}
\title{getExperiment}
\usage{
getExperiment(object)

\S4method{getExperiment}{cignaturesCN}(object)
}
\arguments{
\item{object}{cignaturesCN object}
}
\value{
A cignaturesExpData class object
}
\description{
Extracts and returns copy number features from copy number profiles in a cignaturesCN object.
}
