% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getSampleByComponent.R
\docType{methods}
\name{getsampleByComponent}
\alias{getsampleByComponent}
\alias{getsampleByComponent,cignaturesCN-method}
\title{getsampleByComponent}
\usage{
getsampleByComponent(object)

\S4method{getsampleByComponent}{cignaturesCN}(object)
}
\arguments{
\item{object}{cignaturesCN object}
}
\value{
matrix containing the sample-by-component data
}
\description{
getsampleByComponent
}
