% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getSamples.R
\docType{methods}
\name{getSamples}
\alias{getSamples}
\alias{getSamples,cignaturesCN-method}
\title{getSamples}
\usage{
getSamples(object)

\S4method{getSamples}{cignaturesCN}(object)
}
\arguments{
\item{object}{cignaturesCN object}
}
\value{
A character vector
}
\description{
Extracts sample names from a cignaturesCN object.
}
