% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mac_component_parameters}
\alias{mac_component_parameters}
\title{mac_component_parameters}
\format{
A list object of 6 flexmix class objects
}
\usage{
mac_component_parameters
}
\description{
Flexmix model feature component parameters utilised in macintyre et al.
}
\keyword{datasets}
