% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotActivities.R
\name{plotActivities}
\alias{plotActivities}
\title{plotActivities}
\usage{
plotActivities(object)
}
\arguments{
\item{object}{A cignatureSIG class object}
}
\value{
plot
}
\description{
Plot the copy number signature activites for a given cignatures class object
containing copy number signature activities/exposures. Default ordering by
signature 1
}
