% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cignaturesSIG.R
\name{[,cignaturesSIG,numeric,missing,ANY-method}
\alias{[,cignaturesSIG,numeric,missing,ANY-method}
\title{Extract}
\usage{
\S4method{[}{cignaturesSIG,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{object to extract subset}

\item{i}{Indices of subset}

\item{j}{not used}

\item{...}{not used}

\item{drop}{not used}
}
\description{
Extract
}
