% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getSampleByComponent.R
\docType{methods}
\name{getSampleByComponent}
\alias{getSampleByComponent}
\alias{getSampleByComponent,cignaturesCN-method}
\title{getSampleByComponent}
\usage{
getSampleByComponent(object)

\S4method{getSampleByComponent}{cignaturesCN}(object)
}
\arguments{
\item{object}{cignaturesCN object}
}
\value{
matrix containing the sample-by-component data
}
\description{
getSampleByComponent
}
